requirements.txt
Django>=2.0
djangorestframework
armada-sde>=0.2.2
django-environ
